+++ pkg/backend/path_self.go
@@ -2,7 +2,6 @@ package backend
 
 import (
 	"context"
-	"errors"
 	"strings"
 	"time"
 
@@ -10,7 +9,9 @@ import (
 	"github.com/hashicorp/vault/sdk/logical"
 	"github.com/puppetlabs/leg/errmap/pkg/errmap"
 	"github.com/puppetlabs/leg/errmap/pkg/errmark"
+	"github.com/puppetlabs/leg/timeutil/pkg/clockctx"
 	"github.com/puppetlabs/vault-plugin-secrets-oauthapp/v2/pkg/persistence"
+	"github.com/puppetlabs/vault-plugin-secrets-oauthapp/v2/pkg/provider"
 	"golang.org/x/oauth2"
 )
 
@@ -24,8 +25,6 @@ func (b *backend) selfReadOperation(ctx context.Context, req *logical.Request, d
 		expiryDelta,
 	)
 	switch {
-	case errors.Is(err, ErrNotConfigured):
-		return logical.ErrorResponse("not configured"), nil
 	case errmark.Matches(err, errmark.RuleType(&oauth2.RetrieveError{})) || errmark.MarkedUser(err):
 		return logical.ErrorResponse(errmap.Wrap(errmark.MarkShort(err), "client credentials flow failed").Error()), nil
 	case err != nil:
@@ -37,6 +36,7 @@ func (b *backend) selfReadOperation(ctx context.Context, req *logical.Request, d
 	}
 
 	rd := map[string]interface{}{
+		"server":       entry.AuthServerName,
 		"access_token": entry.Token.AccessToken,
 		"type":         entry.Token.Type(),
 	}
@@ -49,23 +49,72 @@ func (b *backend) selfReadOperation(ctx context.Context, req *logical.Request, d
 		rd["extra_data"] = entry.Token.ExtraData
 	}
 
+	if len(entry.Config.TokenURLParams) > 0 {
+		rd["token_url_params"] = entry.Config.TokenURLParams
+	}
+
+	if len(entry.Config.Scopes) > 0 {
+		rd["scopes"] = entry.Config.Scopes
+	}
+
+	if len(entry.Config.ProviderOptions) > 0 {
+		rd["provider_options"] = entry.Config.ProviderOptions
+	}
+
 	resp := &logical.Response{
 		Data: rd,
 	}
 	return resp, nil
 }
 
-func (b *backend) selfDeleteOperation(ctx context.Context, req *logical.Request, data *framework.FieldData) (*logical.Response, error) {
-	err := b.data.Managers(req.Storage).ClientCreds().WithLock(persistence.ClientCredsName(data.Get("name").(string)), func(cm *persistence.LockedClientCredsManager) error {
-		entry, err := cm.ReadClientCredsEntry(ctx)
-		if err != nil || entry == nil || entry.Token == nil {
-			return nil
-		}
+func (b *backend) selfUpdateOperation(ctx context.Context, req *logical.Request, data *framework.FieldData) (*logical.Response, error) {
+	serverName, ok := data.GetOk("server")
+	if !ok {
+		return logical.ErrorResponse("missing server"), nil
+	}
 
-		entry.Token = nil
-		return cm.WriteClientCredsEntry(ctx, entry)
-	})
-	return nil, err
+	ops, put, err := b.getProviderOperations(ctx, req.Storage, persistence.AuthServerName(serverName.(string)), defaultExpiryDelta)
+	if errmark.MarkedUser(err) {
+		return logical.ErrorResponse(errmark.MarkShort(err).Error()), nil
+	} else if err != nil {
+		return nil, err
+	}
+	defer put()
+
+	entry := &persistence.ClientCredsEntry{
+		AuthServerName: serverName.(string),
+	}
+	entry.Config.TokenURLParams = data.Get("token_url_params").(map[string]string)
+	entry.Config.Scopes = data.Get("scopes").([]string)
+	entry.Config.ProviderOptions = data.Get("provider_options").(map[string]string)
+
+	tok, err := ops.Private().ClientCredentials(
+		clockctx.WithClock(ctx, b.clock),
+		provider.WithURLParams(entry.Config.TokenURLParams),
+		provider.WithScopes(entry.Config.Scopes),
+		provider.WithProviderOptions(entry.Config.ProviderOptions),
+	)
+	if errmark.Matches(err, errmark.RuleType(&oauth2.RetrieveError{})) || errmark.MarkedUser(err) {
+		return logical.ErrorResponse(errmap.Wrap(errmark.MarkShort(err), "client credentials flow failed").Error()), nil
+	} else if err != nil {
+		return nil, err
+	}
+
+	entry.Token = tok
+
+	if err := b.data.ClientCreds.Manager(req.Storage).WriteClientCredsEntry(ctx, persistence.ClientCredsName(data.Get("name").(string)), entry); err != nil {
+		return nil, err
+	}
+
+	return nil, nil
+}
+
+func (b *backend) selfDeleteOperation(ctx context.Context, req *logical.Request, data *framework.FieldData) (*logical.Response, error) {
+	if err := b.data.ClientCreds.Manager(req.Storage).DeleteClientCredsEntry(ctx, persistence.ClientCredsName(data.Get("name").(string))); err != nil {
+		return nil, err
+	}
+
+	return nil, nil
 }
 
 const (
@@ -84,6 +133,23 @@ var selfFields = map[string]*framework.FieldSchema{
 		Description: "Minimum remaining seconds to allow when reusing access token.",
 		Query:       true,
 	},
+	// fields for write operation
+	"server": {
+		Type:        framework.TypeString,
+		Description: "The name of the authorization server to use for this credential.",
+	},
+	"token_url_params": {
+		Type:        framework.TypeKVPairs,
+		Description: "Specifies the additional query parameters to add to the token URL.",
+	},
+	"scopes": {
+		Type:        framework.TypeCommaStringSlice,
+		Description: "The scopes to request for authorization.",
+	},
+	"provider_options": {
+		Type:        framework.TypeKVPairs,
+		Description: "Specifies any provider-specific options.",
+	},
 }
 
 const selfHelpSynopsis = `
@@ -91,9 +157,9 @@ Provides access tokens for this application using the OAuth 2.0 client credentia
 `
 
 const selfHelpDescription = `
-This endpoint allows you to read and delete OAuth 2.0 access tokens
-using the client credentials flow. Reads from a given path are cached
-until expiration and will be automatically resubmitted to the IdP as
+This endpoint allows you to manage OAuth 2.0 access tokens using the
+client credentials flow. Reads from a given path are cached until
+expiration and will be automatically resubmitted to the IdP as
 needed.
 `
 
@@ -105,6 +171,10 @@ func pathSelf(b *backend) *framework.Path {
 			logical.ReadOperation: &framework.PathOperation{
 				Callback: b.selfReadOperation,
 				Summary:  "Get a current access token for this credential.",
+			},
+			logical.UpdateOperation: &framework.PathOperation{
+				Callback: b.selfUpdateOperation,
+				Summary:  "Write a new credential or update an existing credential.",
 			},
 			logical.DeleteOperation: &framework.PathOperation{
 				Callback: b.selfDeleteOperation,