+++ pkg/backend/token_clientcreds.go
@@ -12,36 +12,32 @@ import (
 
 func (b *backend) updateClientCredsToken(ctx context.Context, storage logical.Storage, keyer persistence.ClientCredsKeyer, expiryDelta time.Duration) (*persistence.ClientCredsEntry, error) {
 	var entry *persistence.ClientCredsEntry
-	err := b.data.Managers(storage).ClientCreds().WithLock(keyer, func(cm *persistence.LockedClientCredsManager) error {
+	err := b.data.ClientCreds.WithLock(keyer, func(ch *persistence.LockedClientCredsHolder) error {
+		cm := ch.Manager(storage)
+
 		// In case someone else updated this token from under us, we'll re-request
 		// it here with the lock acquired.
 		candidate, err := cm.ReadClientCredsEntry(ctx)
 		switch {
-		case err != nil:
+		case err != nil || candidate == nil:
 			return err
-		case candidate == nil:
-			candidate = &persistence.ClientCredsEntry{}
 		case b.tokenValid(candidate.Token, expiryDelta):
 			entry = candidate
 			return nil
 		}
 
-		c, err := b.getCache(ctx, storage)
+		ops, put, err := b.getProviderOperations(ctx, storage, persistence.AuthServerName(candidate.AuthServerName), expiryDelta)
 		if err != nil {
 			return err
-		} else if c == nil {
-			return ErrNotConfigured
 		}
+		defer put()
 
-		updated, err := c.
-			ProviderWithTimeout(expiryDelta).
-			Private(c.Config.ClientID, c.Config.ClientSecret).
-			ClientCredentials(
-				clockctx.WithClock(ctx, b.clock),
-				provider.WithURLParams(candidate.Config.TokenURLParams),
-				provider.WithScopes(candidate.Config.Scopes),
-				provider.WithProviderOptions(candidate.Config.ProviderOptions),
-			)
+		updated, err := ops.Private().ClientCredentials(
+			clockctx.WithClock(ctx, b.clock),
+			provider.WithURLParams(candidate.Config.TokenURLParams),
+			provider.WithScopes(candidate.Config.Scopes),
+			provider.WithProviderOptions(candidate.Config.ProviderOptions),
+		)
 		if err != nil {
 			return err
 		}
@@ -60,7 +56,7 @@ func (b *backend) updateClientCredsToken(ctx context.Context, storage logical.St
 }
 
 func (b *backend) getUpdateClientCredsToken(ctx context.Context, storage logical.Storage, keyer persistence.ClientCredsKeyer, expiryDelta time.Duration) (*persistence.ClientCredsEntry, error) {
-	entry, err := b.data.Managers(storage).ClientCreds().ReadClientCredsEntry(ctx, keyer)
+	entry, err := b.data.ClientCreds.Manager(storage).ReadClientCredsEntry(ctx, keyer)
 	switch {
 	case err != nil:
 		return nil, err