+++ pkg/provider/basic.go
@@ -116,6 +116,10 @@ func (bo *basicOperations) DeviceCodeExchange(ctx context.Context, deviceCode st
 }
 
 func (bo *basicOperations) AuthCodeExchange(ctx context.Context, code string, opts ...AuthCodeExchangeOption) (*Token, error) {
+	if bo.clientSecret == "" {
+		return nil, errmark.MarkUser(ErrMissingClientSecret)
+	}
+
 	o := &AuthCodeExchangeOptions{}
 	o.ApplyOptions(opts)
 
@@ -170,6 +174,10 @@ func (bo *basicOperations) RefreshToken(ctx context.Context, t *Token, opts ...R
 }
 
 func (bo *basicOperations) ClientCredentials(ctx context.Context, opts ...ClientCredentialsOption) (*Token, error) {
+	if bo.clientSecret == "" {
+		return nil, errmark.MarkUser(ErrMissingClientSecret)
+	}
+
 	o := &ClientCredentialsOptions{}
 	o.ApplyOptions(opts)